Inital commit, adding unit tests for models
